Completed version of Scraper
This is the completed webapp version of the Scraper. This allows dynamic input from the user making it so you don't have to go into the code in order to change what you want request. It displays all scraped data on the website with the active links to visit the product listed.